Area-Proportional Euler Diagrams (deprecated)

Note: This function has been deprecated; please use euler instead.

eulerr(sets, ...)

# S3 method for default
eulerr(sets, cost = NULL, ...)

# S3 method for matrix
eulerr(sets, by = NULL, cost = NULL, ...)

# S3 method for data.frame
eulerr(sets, by = NULL, cost = NULL, ...)

Arguments

sets

Set relationships as a named numeric vector, matrix, or data.frame. (See the methods (by class) section for details.)

…

Currently ignored.

cost

Deprecated.

by

A factor or character vector used in by to split the data.frame or matrix and compute euler diagrams for each split.

Value

A list object of class 'eulerr' with the following parameters.

Details

Compute Euler diagrams (a generalization of Venn diagrams) using numerical optimization to find exact or approximate solutions to a specification of set relationships.

If by is specified, eulerr returns a list of euler diagrams that can be plotted in facets via a special plot method.

The fit is optimized using either the cost function used in the eulerAPE software package or the stress statistic of the R package venneuler. The eulerAPE cost function is defined as

$$% \frac{1}{n} \sum_{i=1}^{n} \frac{(y_i - \hat{y}_i) ^ 2}{\hat{y}_i} $$

where \(\hat{y}\) are the estimates of \(y\) that are currently explored during optimization. For venneuler, the stress function is defined as

$$% \frac{\sum_{i=1}^{n} (y_i - \hat{y}_i) ^ 2}{\sum_{i=1}^{n} y_i ^ 2} $$

where \(\hat{y}\) are OLS estimates from the regression of the fitted areas on the original areas that are currently being explored during optimization.

Methods (by class)

default: A named numeric vector, with interactions seperated by an ampersand, for instance `A&B` = 10. Missing interactions are treated as being 0.

matrix: A matrix of logical vectors with columns representing sets and rows representing each observation's set relationships (see examples).

data.frame: A data.frame that can be converted to a matrix of logicals (as in the description above) via as.matrix.

References

Wilkinson L. Exact and Approximate Area-Proportional Circular Venn and Euler Diagrams. IEEE Transactions on Visualization and Computer Graphics [Internet]. 2012 Feb [cited 2016 Apr 9];18(2):321–31. Available from: http://doi.org/10.1109/TVCG.2011.56

Micallef L, Rodgers P. eulerAPE: Drawing Area-Proportional 3-Venn Diagrams Using Ellipses. PLOS ONE [Internet]. 2014 Jul [cited 2016 Dec 10];9(7):e101717. Available from: http://dx.doi.org/10.1371/journal.pone.0101717