Deprecated functions in package eulerr.
Source: R/eulerr_deprecated.R

The functions listed below are deprecated and will be defunct in the near future. When possible, alternative functions with similar functionality are also mentioned. Help pages for deprecated functions are available at help("_deprecated").

prepanel.euler(x, y, ra, rb, phi, subscripts, ...)

panel.euler(x, y, ra, rb, phi, subscripts, fill = superpose.polygon$col,
  lty = superpose.polygon$lty, lwd = superpose.polygon$lwd,
  border = superpose.polygon$border, alpha = superpose.polygon$alpha,
  fill_alpha = 0.4, fontface = "bold", quantities = FALSE,
  labels = NULL, original.values, fitted.values, mode = c("split",
  "overlay"), ...)

panel.euler.circles(x, y, r, border = "black", fill = "transparent",
  fill_alpha = 0.4, ..., identifier = NULL, name.type = "panel", col,
  font, fontface)

panel.euler.ellipses(x, y, ra, rb, phi, fill = "transparent",
  fill_alpha = 0.4, border = "black", mode = c("split", "overlay"),
  identifier = NULL, n = 200, fitted.values, ..., name.type = "panel",
  col, font, fontface)

panel.euler.labels(x, y, ra, rb, phi, labels, quantities = FALSE,
  original.values, fitted.values, ...)

label(x, labels = NULL)

# S3 method for euler
label(x, labels = NULL)

prepanel.euler

For prepanel.euler(), use plot.euler().

panel.euler

For panel.euler(), use plot.euler().

panel.euler.circles

For panel.euler.circles(), use plot.euler().

panel.euler.ellipses

For panel.euler.ellipses(), use plot.euler().

panel.euler.labels

For panel.euler.labels(), use plot.euler().

label.euler

For label.euler(), use plot.euler().