Fruits

A synthethic data set of preferences for fruits and their overlaps, generated only to be a showcase for the examples for this package.

fruits

Format

A data.frame with 100 observations of 5 variables:

banana

whether the person likes bananas, a logical

apple

whether the person likes apples, a logical

orange

whether the person likes oranges, a logical

sex

the sex of the person, a factor with levels 'male' and 'female'

age

the age of the person, a factor with levels 'child' and 'adult'