Panel Function for Euler Diagrams

Panel Function for Euler Diagrams

panel.euler(x, y, r, subscripts, fill = superpose.polygon$col,
  lty = superpose.polygon$lty, lwd = superpose.polygon$lwd,
  border = superpose.polygon$border, alpha = superpose.polygon$alpha,
  fontface = "bold", counts = TRUE, labels = NULL, original.values,
  fitted.values, ...)

Arguments

x
X coordinates for the circle centers.

y
Y coordinates for the circle centers.

r
Radii.

subscripts
A vector of subscripts (See lattice::xyplot()).

fill
Fill color for circles. (See grid::gpar().)

lty
Line type for circles. (See grid::gpar().)

lwd
Line weight for circles. (See grid::gpar().)

border
Border color for circles.

alpha
Alpha for circles. Note that plot.euler() by default modifies the alpha of col instead to avoid affecting the alpha of the borders. (See grid::gpar().)

fontface
Fontface for the labels. (See grid::gpar().)

counts
Plots the original values for the disjoint set combinations (original.values). Can also be a list, in which the contents of the list will be passed on to lattice::panel.text() to modify the appearance of the counts.

labels
Labels to plot on the circles.

original.values
Original values for the disjoint set combinations.

fitted.values
Fitted values for the disjoint set combinations.

...
Passed down to panel.euler.circles() and panel.euler.labels().

Value

Plots euler diagrams inside a trellis panel.

See also

grid::gpar().