Panel Function for Euler Ellipses (deprecated)
Source: R/eulerr_deprecated.R

Panel Function for Euler Ellipses (deprecated)

panel.euler.ellipses(x, y, ra, rb, phi, fill = "transparent",
  fill_alpha = 0.4, border = "black", mode = c("split", "overlay"),
  identifier = NULL, n = 200, fitted.values, ..., name.type = "panel",
  col, font, fontface)

Arguments

x
X coordinates for the centers.

y
Y coordinates for the centers.

ra
Semi-major axes.

rb
Semi-minor axes.

phi
Rotation of the ellipse (as the counter-clockwise angle from the positive x-axis to the semi-major axis).

border
Border color.

fill
Ellipse fill.

n
Number of vertices to draw for each ellipse.

...
Passed on to grid::grid.polygon().

col
Ignored

font
Ignored

fontface
Ignored

identifier
A character string that is prepended to the name of the grob that is created.

name.type
A character value indicating whether the name of the grob should have panel or strip information added to it. Typically either "panel", "strip", "strip.left", or "" (for no extra information).

mode
"split", the default, splits up the diagram into individual polygons and blends the colors of the overlapping shapes using color averaging in the CIELAB color space. "overlay" superposes sets and should be used in conjunction with a suitable fill_alpha value.

original.values
Original values for the disjoint set combinations.

fitted.values
Fitted values for the disjoint set combinations.

Value

Plots ellipses inside a trellis panel.

See also