Print (plot) Euler diagram

This function is responsible for the actual drawing of 'eulergram' objects created through plot.euler(). print.eulergram() is an alias for plot.eulergram(), which has been provided so that plot.euler() gets called automatically.

Usage

# S3 method for class 'eulergram'
plot(x, newpage = TRUE, ...)

# S3 method for class 'eulergram'
print(x, ...)

Arguments

x

an object of class 'eulergram', usually the output of plot.euler()

newpage

if TRUE, opens a new page via grid.newpage() to draw on

...

ignored

Value

A plot is drawn on the current device using grid::Grid() graphics.