Print Euler diagram fits
Source: R/print.euler.R

This function is responsible for printing fits from euler() and provides a summary of the fit. Prints a data frame of the original set relationships and the fitted values as well as diagError and stress statistics.

# S3 method for euler
print(x, round = 3, vsep = strrep("-", 0.75 *
  getOption("width")), ...)

Arguments

x
'euler' object from euler()

round
number of decimal places to round to

vsep
character string to paste in between euler objects when x is a nested euler object

...
arguments passed to base::print.data.frame()

Value

Summary statistics of the fitted Euler diagram is printed to screen or a plot is generated.

See also