Area-Proportional Euler Diagrams

Fit euler diagrams (a generalization of venn diagrams) using numerical optimization to find exact or approximate solutions to a specification of set relationships.

euler(combinations, ...)

# S3 method for default
euler(combinations, input = c("disjoint", "union"), ...)

# S3 method for matrix
euler(combinations, by = NULL, ...)

# S3 method for data.frame
euler(combinations, by = NULL, ...)

Arguments

combinations

Set relationships as a named numeric vector, matrix, or data.frame. (See the methods (by class) section for details.)

…

Currently ignored.

input

The type of input: disjoint class combinations (disjoint) or unions (union).

by

A factor or character matrix to be used in by to split the data.frame or matrix of set combinations.

Value

A list object of class 'euler' with the following parameters.

Details

If by is specified, euler returns a list of diagrams separated by the categorical variables in by.

The fit minimizes the sums of squared residuals between the areas in the euler diagram and the user's initial specification (as disjoint class combinations),

$$\sum_{i=1}^{n} (y_i - \hat{y}_i) ^ 2 $$

where \(\hat{y}\) are the estimates of \(y\) that are explored during optimization.

Diagnostics are provided by the fit as the stress statistic from venneuler:

$$ \frac{\sum_{i=1}^{n} (y_i - \hat{y}_i) ^ 2}{\sum_{i=1}^{n} y_i ^ 2} $$

where \(\hat{y}\) are OLS estimates from the regression of the fitted areas on the original areas that are currently being explored during optimization.

We also return diag_error and region_error from eulerAPE. region_error is computed as

$$ \left| \frac{y_i}{\sum y_i} - \frac{\hat{y}_i}{\sum \hat{y}_i}\right| $$

whereas diagError is the maximum of regionError.

Methods (by class)

default: A named numeric vector, with interactions seperated by an ampersand, for instance `A&B` = 10. Missing interactions are treated as being 0.

matrix: A matrix of logical vectors with columns representing sets and rows representing each observation's set relationships (see examples).

data.frame: A data.frame that can be converted to a matrix of logicals (as in the description above) via as.matrix.

References

Wilkinson L. Exact and Approximate Area-Proportional Circular Venn and Euler Diagrams. IEEE Transactions on Visualization and Computer Graphics [Internet]. 2012 Feb [cited 2016 Apr 9];18(2):321–31. Available from: http://doi.org/10.1109/TVCG.2011.56

Micallef L, Rodgers P. eulerAPE: Drawing Area-Proportional 3-Venn Diagrams Using Ellipses. PLOS ONE [Internet]. 2014 Jul [cited 2016 Dec 10];9(7):e101717. Available from: http://dx.doi.org/10.1371/journal.pone.0101717

See also

plot.euler, print.euler

Examples

fit1 <- euler(c("A" = 1, "B" = 0.4, "C" = 3, "A&B" = 0.2))

# Same result as above
fit2 <- euler(c("A" = 1, "B" = 0.4, "C" = 3,
                "A&B" = 0.2, "A&C" = 0, "B&C" = 0,
                "A&B&C" = 0))

# Using the matrix method
mat <- cbind(A = sample(c(TRUE, TRUE, FALSE), size = 50, replace = TRUE),
             B = sample(c(TRUE, FALSE), size = 50, replace = TRUE))
fit3 <- euler(mat)

# Using grouping via the 'by' argument
dat <- data.frame(
  A = sample(c(TRUE, FALSE), size = 100, replace = TRUE),
  B = sample(c(TRUE, TRUE, FALSE), size = 100, replace = TRUE),
  gender = sample(c("Men", "Women"), size = 100, replace = TRUE),
  nation = sample(c("Sweden", "Denmark"), size = 100, replace = TRUE)
)

fit4 <- euler(dat[, 1:2], by = dat[, 3:4])

# A set with no perfect solution
euler(c("a" = 3491, "b" = 3409, "c" = 3503,
        "a&b" = 120, "a&c" = 114, "b&c" = 132,
        "a&b&c" = 50))

#>       original   fitted residuals region_error
#> a         3491 3490.993     0.007        0.001
#> b         3409 3408.994     0.006        0.001
#> c         3503 3502.991     0.009        0.002
#> a&b        120  119.972     0.028        0.000
#> a&c        114  113.999     0.001        0.000
#> b&c        132  131.992     0.008        0.000
#> a&b&c       50    0.000    50.000        0.005
#> 
#> diag_error:  0.005 
#> stress:      0