Get or set global graphical parameters for eulerr

This function provides a means to set default parameters for functions in eulerr. Query eulerr_options() (without any argument) to see all the available options and read more about the plot-related ones in grid::gpar() and graphics::par().

eulerr_options(...)

Arguments

...
objects to update the global graphical parameters for eulerr with.

Value

This function gets or sets updates in the global environment that are used in plot.euler().

Details

Currently, the following items will be considered:

pointsize

size in pts to be used as basis for fontsizes and some margin sizes in the resulting plot

fills

a list of items fill and alpha

edges

a list of items col, alpha, lex, lwd, and lty

labels

a list of items rot, col, alpha, fontsize, cex, fontfamily, fontface, lineheight, and font

quantities

a list of items rot, col, alpha, fontsize, cex, fontfamily, lineheight, and font

strips

col, alpha, fontsize, cex, fontfamily, lineheight, and font

legend

arguments to grid::legendGrob() as well as col, alpha, fontsize, cex, fontfamily, lineheight, and font

main

arguments to grid::textGrob()

See also

Examples

eulerr_options(edges = list(col = "blue"), fontsize = 10)
eulerr_options(n_threads = 2)